pyproject.toml
[tool.poetry]
name = "tap-cassandra"
version = "0.0.1"
description = "`tap-cassandra` is a Singer tap for Cassandra, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Data arts <developers@datarts.tech>"]
keywords = [
    "ELT",
    "Cassandra",
]
license = "MIT"

[tool.poetry.dependencies]
python = "<3.12,>=3.7.1"
singer-sdk = { version="^0.34.1" }
fs-s3fs = { version = "^1.1.1", optional = true }
cassandra-driver = "^3.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
singer-sdk = { version="^0.34.1", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true

[tool.ruff]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
]
select = ["ALL"]
src = ["tap_cassandra"]
target-version = "py37"


[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["tap_cassandra"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-cassandra = 'tap_cassandra.tap:TapCassandra.cli'
